Add ability for button to be classed "olorinPlugin"
Right now the plugin only looks for a button with the id "webPrint". The problem is, if there are multiple buttons with that id, only the first button will work. This commit adds the ability for the buttons to use a class instead. The class name is "olorinPlugin" which is unlikely to clash with existing class names.